+++ src/models/umx.py
@@ -305,254 +305,6 @@ class OpenUnmix(nn.Module):
         
         return _num_parameters
 
-"""
-CrossNet-Open-Unmix
-    Reference: "All for One and One for All: Improving Music Separation by Bridging Networks"
-    See https://arxiv.org/abs/2010.04228
-"""
-class CrossNetOpenUnmix(nn.Module):
-    def __init__(self, in_channels, hidden_channels=512, num_layers=3, n_bins=None, max_bin=None, dropout=None, causal=False, rnn_type='lstm', bridge=True, sources=__sources__, eps=EPS):
-        """
-        Args:
-            in_channels <int>: Input channels
-            hidden_channels <int>: Hidden channels in LSTM
-            num_layers <int>: # of LSTM layers
-            n_bins <int>: # of frequency bins
-            max_bin <int>: If none, max_bin = n_bins
-            dropout <float>: Dropout rate in LSTM
-            causal <bool>: Causality
-            sources <list<str>>: Target sources
-            eps <float>: Small value for numerical stability
-        """
-        super().__init__()
-        
-        net = {}
-        
-        for source in sources:
-            net[source] = OpenUnmix(in_channels, hidden_channels, num_layers=num_layers, n_bins=n_bins, max_bin=max_bin, dropout=dropout, causal=causal, rnn_type=rnn_type, eps=eps)
-
-        self.backbone = nn.ModuleDict(net)
-
-        # Hyperparameters
-        self.in_channels, self.n_bins = in_channels, n_bins
-        self.hidden_channels, self.out_channels = hidden_channels, hidden_channels
-        self.num_layers = num_layers
-        self.max_bin = max_bin
-
-        self.dropout = dropout
-        self.causal = causal
-        self.rnn_type = rnn_type
-        self.bridge = bridge
-
-        self.sources = sources
-
-        self.eps = eps
-        
-    def forward(self, input):
-        """
-        Args:
-            input <torch.Tensor>: (batch_size, 1, in_channels, n_bins, n_frames)
-        Returns:
-            output <torch.Tensor>: (batch_size, n_sources, in_channels, n_bins, n_frames)
-        """
-        n_bins, max_bin = self.n_bins, self.max_bin
-
-        input = input.squeeze(dim=1)
-
-        if max_bin == n_bins:
-            x_valid = input
-        else:
-            sections = [max_bin, n_bins - max_bin]
-            x_valid, _ = torch.split(input, sections, dim=2)
-
-        if self.bridge:
-            output = self.forward_bridge(input, x_valid)
-        else:
-            output = self.forward_no_bridge(input, x_valid)
-        
-        return output
-
-    def forward_no_bridge(self, input, x_valid):
-        n_bins, max_bin = self.n_bins, self.max_bin
-        in_channels, hidden_channels, out_channels = self.in_channels, self.hidden_channels, self.out_channels
-        eps = self.eps
-
-        batch_size, _, _, n_frames = x_valid.size()
-
-        x_sources = []
-
-        for source in self.sources:
-            x_source = (x_valid - self.backbone[source].bias_in.unsqueeze(dim=1)) / (torch.abs(self.backbone[source].scale_in.unsqueeze(dim=1)) + eps) # (batch_size, n_channels, max_bin, n_frames)
-            x_source = x_source.permute(0, 3, 1, 2).contiguous() # (batch_size, n_frames, n_channels, max_bin)
-            x_source = x_source.view(batch_size * n_frames, in_channels * max_bin)
-            x_source = self.backbone[source].block(x_source) # (batch_size * n_frames, hidden_channels)
-            x_source = x_source.view(batch_size, n_frames, hidden_channels)
-            x_sources.append(x_source)
-        
-        x = torch.stack(x_sources, dim=0) # (n_sources, batch_size, n_frames, hidden_channels)
-        x_sources = []
-
-        for idx, source in enumerate(self.sources):
-            x_source = x[idx]
-            x_source_lstm, _ = self.backbone[source].rnn(x_source) # (batch_size, n_frames, out_channels)
-            x_source = torch.cat([x_source, x_source_lstm], dim=2) # (batch_size, n_frames, hidden_channels + out_channels)
-            x_source = x_source.view(batch_size * n_frames, hidden_channels + out_channels)
-            x_sources.append(x_source)
-        
-        x = torch.stack(x_sources, dim=0) # (n_sources, batch_size * n_frames, hidden_channels + out_channels)
-        output = []
-
-        for source in self.sources:
-            x_source = x[idx]
-            x_source_full = self.backbone[source].net(x_source) # (batch_size * n_frames, n_bins)
-            x_source_full = x_source_full.view(batch_size, n_frames, in_channels, n_bins)
-            x_source_full = x_source_full.permute(0, 2, 3, 1).contiguous() # (batch_size, in_channels, n_bins, n_frames)
-            x_source_full = self.backbone[source].scale_out.unsqueeze(dim=1) * x_source_full + self.backbone[source].bias_out.unsqueeze(dim=1)
-            x_source_full = self.backbone[source].relu2d(x_source_full)
-            x_source = x_source_full * input
-            output.append(x_source)
-        
-        output = torch.stack(output, dim=1) # (batch_size, n_sources, in_channels, n_bins, n_frames)
-
-        return output
-    
-    def forward_bridge(self, input, x_valid):
-        n_bins, max_bin = self.n_bins, self.max_bin
-        in_channels, hidden_channels, out_channels = self.in_channels, self.hidden_channels, self.out_channels
-        eps = self.eps
-
-        batch_size, _, _, n_frames = x_valid.size()
-
-        x_sum = []
-
-        for source in self.sources:
-            x_source = (x_valid - self.backbone[source].bias_in.unsqueeze(dim=1)) / (torch.abs(self.backbone[source].scale_in.unsqueeze(dim=1)) + eps) # (batch_size, n_channels, max_bin, n_frames)
-            x_source = x_source.permute(0, 3, 1, 2).contiguous() # (batch_size, n_frames, n_channels, max_bin)
-            x_source = x_source.view(batch_size * n_frames, in_channels * max_bin)
-            x_source = self.backbone[source].block(x_source) # (batch_size * n_frames, hidden_channels)
-            x_source = x_source.view(batch_size, n_frames, hidden_channels)
-            x_sum.append(x_source)
-        
-        x_sum = torch.stack(x_sum, dim=0) # (n_sources, batch_size, n_frames, hidden_channels)
-        x = x_sum.mean(dim=0) # (batch_size, n_frames, hidden_channels)
-        x_sum = []
-
-        for source in self.sources:
-            x_source_lstm, _ = self.backbone[source].rnn(x) # (batch_size, n_frames, out_channels)
-            x_source = torch.cat([x, x_source_lstm], dim=2) # (batch_size, n_frames, hidden_channels + out_channels)
-            x_source = x_source.view(batch_size * n_frames, hidden_channels + out_channels)
-            x_sum.append(x_source)
-        
-        x_sum = torch.stack(x_sum, dim=0) # (n_sources, batch_size * n_frames, hidden_channels + out_channels)
-        x = x_sum.mean(dim=0) # (batch_size * n_frames, hidden_channels + out_channels)
-        output = []
-
-        for source in self.sources:
-            x_source_full = self.backbone[source].net(x) # (batch_size * n_frames, n_bins)
-            x_source_full = x_source_full.view(batch_size, n_frames, in_channels, n_bins)
-            x_source_full = x_source_full.permute(0, 2, 3, 1).contiguous() # (batch_size, in_channels, n_bins, n_frames)
-            x_source_full = self.backbone[source].scale_out.unsqueeze(dim=1) * x_source_full + self.backbone[source].bias_out.unsqueeze(dim=1)
-            x_source_full = self.backbone[source].relu2d(x_source_full)
-            x_source = x_source_full * input
-            output.append(x_source)
-        
-        output = torch.stack(output, dim=1) # (batch_size, n_sources, in_channels, n_bins, n_frames)
-
-        return output
-    
-    def get_config(self):
-        config = {
-            'in_channels': self.in_channels,
-            'hidden_channels': self.hidden_channels,
-            'num_layers': self.num_layers,
-            'n_bins': self.n_bins,
-            'max_bin': self.max_bin,
-            'dropout': self.dropout,
-            'causal': self.causal,
-            'rnn_type': self.rnn_type,
-            'bridge': self.bridge,
-            'sources': self.sources,
-            'eps': self.eps
-        }
-        
-        return config
-    
-    @classmethod
-    def build_from_config(cls, config_path):
-        with open(config_path, 'r') as f:
-            config = yaml.safe_load(f)
-
-        in_channels = config['in_channels']
-
-        hidden_channels = config['hidden_channels']
-        num_layers = config['num_layers']
-        n_bins, max_bin = config['n_bins'], config['max_bin']
-        dropout = config['dropout']
-        causal = config['causal']
-        rnn_type = config['rnn_type']
-        bridge = config['bridge']
-
-        sources = config['sources']
-
-        eps = config.get('eps') or EPS
-
-        model = cls(
-            in_channels,
-            hidden_channels=hidden_channels,
-            num_layers=num_layers,
-            n_bins=n_bins, max_bin=max_bin,
-            dropout=dropout,
-            causal=causal,
-            sources=sources,
-            rnn_type=rnn_type,
-            bridge=bridge,
-            eps=eps
-        )
-        
-        return model
-    
-    @classmethod
-    def build_model(cls, model_path):
-        config = torch.load(model_path, map_location=lambda storage, loc: storage)
-    
-        in_channels = config['in_channels']
-        hidden_channels = config['hidden_channels']
-        num_layers = config['num_layers']
-        n_bins, max_bin = config['n_bins'], config['max_bin']
-        dropout = config['dropout']
-        causal = config['causal']
-        rnn_type = config['rnn_type']
-        bridge = config['bridge']
-
-        sources = config['sources']
-
-        eps = config.get('eps') or EPS
-        
-        model = cls(
-            in_channels,
-            hidden_channels=hidden_channels,
-            num_layers=num_layers,
-            n_bins=n_bins, max_bin=max_bin,
-            dropout=dropout,
-            causal=causal,
-            rnn_type=rnn_type,
-            bridge=bridge,
-            sources=sources,
-            eps=eps
-        )
-        
-        return model
-    
-    @property
-    def num_parameters(self):
-        _num_parameters = 0
-        
-        for p in self.parameters():
-            if p.requires_grad:
-                _num_parameters += p.numel()
-                
-        return _num_parameters
-
 class TransformBlock1d(nn.Module):
     def __init__(self, in_channels, out_channels, bias=True, nonlinear=None, eps=EPS):
         super().__init__()
@@ -602,30 +354,8 @@ def _test_openunmix():
     print(model)
     print(input.size(), output.size())
 
-def _test_crossnet_openunmix():
-    batch_size = 6
-    in_channels = 2
-    n_bins, max_bin = 2049, 1487
-    n_frames = 100
-    dropout = 0.4
-
-    input = torch.randn(batch_size, 1, in_channels, n_bins, n_frames)
-
-    print('-'*10, "Non causal", '-'*10)
-    causal = False
-    model = CrossNetOpenUnmix(in_channels=in_channels, n_bins=n_bins, max_bin=max_bin, dropout=dropout, causal=causal)
-    output = model(input)
-
-    print(model)
-    print(model.num_parameters)
-    print(input.size(), output.size())
-
 if __name__ == '__main__':
     torch.manual_seed(111)
 
     print("="*10, "Open-Unmix (UMX)", "="*10)
-    # _test_openunmix()
-    print()
-
-    print("="*10, "Cross-Net Open-Unmix (X-UMX)", "="*10)
-    _test_crossnet_openunmix()
+    _test_openunmix()
